refactor(editor): Move formatUserDisplayName into workflowReviews.utils (no-changelog)
